spy_quit(): rv is a LONG not int
The problem occurs on 64-bits systems with sign extension from
0x80100016 to 0xFFFFFFFF80100016


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@5970 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/spy/pcsc-spy.c
@@ -171,7 +171,7 @@ static void spy_enter(const char *fname)
     spy_line(">|%d|%d|%s", profile_time.tv_sec, profile_time.tv_usec, fname);
 }
 
-static void spy_quit(const char *fname, int rv)
+static void spy_quit(const char *fname, LONG rv)
 {
     struct timeval profile_time;
